refactor: Improve dialog close animation and add robust timeout handling
- Extract DIALOG_ANIMATION_DURATION as a constant for maintainability
- Add useRef to track and cleanup timeouts preventing memory leaks
- Add useEffect cleanup to clear timeouts on component unmount
- Move DIALOG_ANIMATION_DURATION to constants.ts for centralized configuration
- Move validator tests to __test__ directory for better organization
- Fix TypeScript typing for useRef with explicit undefined type

These improvements address code review feedback and make the form reset
behavior more robust and maintainable.
