+++ src/components/addresses/create-address-dialog.tsx
@@ -8,7 +8,7 @@ import {
 } from "@tabler/icons-react";
 import { useMutation } from "convex/react";
 import { ConvexError } from "convex/values";
-import { useMemo, useState } from "react";
+import { useEffect, useMemo, useRef, useState } from "react";
 import { useForm } from "react-hook-form";
 import { toast } from "sonner";
 import { z } from "zod";
@@ -37,6 +37,7 @@ import {
 import { Input } from "@/components/ui/input";
 import { api } from "@/convex/_generated/api";
 import {
+  DIALOG_ANIMATION_DURATION,
   getTokenNetworkInfo,
   getValidNetworksForToken,
   isValidTokenNetworkCombination,
@@ -113,6 +114,7 @@ const formSchema = z
 export function CreateAddressDialog() {
   const [open, setOpen] = useState(false);
   const addAddress = useMutation(api.addresses.add);
+  const resetTimeoutRef = useRef<NodeJS.Timeout | undefined>(undefined);
 
   const form = useForm<z.infer<typeof formSchema>>({
     resolver: zodResolver(formSchema),
@@ -223,12 +225,25 @@ export function CreateAddressDialog() {
   const handleOpenChange = (newOpen: boolean) => {
     setOpen(newOpen);
     if (!newOpen) {
+      // Clear any existing timeout
+      if (resetTimeoutRef.current) {
+        clearTimeout(resetTimeoutRef.current);
+      }
       // Delay form reset to prevent visual flicker during close animation
-      setTimeout(() => {
+      resetTimeoutRef.current = setTimeout(() => {
         form.reset();
-      }, 300); // Match dialog animation duration
+      }, DIALOG_ANIMATION_DURATION);
     }
   };
+
+  // Cleanup timeout on unmount
+  useEffect(() => {
+    return () => {
+      if (resetTimeoutRef.current) {
+        clearTimeout(resetTimeoutRef.current);
+      }
+    };
+  }, []);
 
   // Handle regenerate code button click
   const handleRegenerateCode = () => {
